% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{Driver}
\alias{Driver}
\title{Driver}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Driver
}
\description{
Driver is the class that provides access to the bolt protocol
}
\keyword{datasets}
