% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{GraphDatabase}
\alias{GraphDatabase}
\title{GraphDatabase}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GraphDatabase
}
\description{
Accessor for :class:`.Driver` construction.
}
\examples{
driver <- GraphDatabase$new()$driver(uri='bolt://localhost:7687',user_id='neo4j',password='neo4j')
}
\keyword{datasets}
