% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{Session}
\alias{Session}
\title{Session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Session
}
\description{
Session is the class that allows transactions in the Neo4jDB
}
\keyword{datasets}
