% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{StatementResult}
\alias{StatementResult}
\title{StatementResult}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StatementResult
}
\description{
StatementResult is the class that is the return from execution Neo4jDB
}
\keyword{datasets}
