% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{Transaction}
\alias{Transaction}
\title{Transaction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Transaction
}
\description{
Transaction is the class that allows execution in the Neo4jDB
}
\keyword{datasets}
