% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{GraphDatabase}
\alias{GraphDatabase}
\title{GraphDatabase}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GraphDatabase
}
\description{
Accessor for :class:`.Driver` construction.
}
\examples{
uri <-  Sys.getenv("NEO4J_URI")
uid <- Sys.getenv("NEO4J_USER")
pswd = Sys.getenv("NEO4J_PSWD")
driver <- GraphDatabase$new()$driver(uri=uri,user_id=uid,password=pswd)
}
\keyword{datasets}
