% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_database.R
\docType{data}
\name{Session}
\alias{Session}
\title{Session}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Session
}
\description{
S A :class:`.Session` is a logical context for transactional units
of work. Connections are drawn from the :class:`.Driver` connection
pool as required.
Session creation is a lightweight operation and sessions are not thread
safe. Therefore a session should generally be short-lived, and not
span multiple threads.
}
\keyword{datasets}
